+++ common/src/main/java/fr/loudo/narrativecraft/client/editors/interaction/ClientInteractionMakerEditorMaker.java
@@ -130,7 +130,7 @@ public class ClientInteractionMakerEditorMaker implements EditorMaker {
     }
 
     @Override
-    public void stop() {}
+    public void close() {}
 
     @Override
     public void tick() {}
@@ -337,7 +337,7 @@ public class ClientInteractionMakerEditorMaker implements EditorMaker {
     }
 
     public void quit() {
-        playerSession.setEditor(null);
+        playerSession.requestEditorClose();
         minecraft.gui.setScreen(null);
     }