+++ common/src/main/java/fr/loudo/narrativecraft/client/screens/mainScreen/MainScreen.java
@@ -27,8 +27,6 @@ import com.mojang.blaze3d.platform.NativeImage;
 import fr.loudo.narrativecraft.NarrativeCraftMod;
 import fr.loudo.narrativecraft.client.ClientNarrativeCraftMod;
 import fr.loudo.narrativecraft.client.session.ClientPlayerSession;
-import fr.loudo.narrativecraft.editors.EditorMaker;
-import fr.loudo.narrativecraft.network.BiStopEditorMaker;
 import fr.loudo.narrativecraft.network.story.C2SPlayStory;
 import fr.loudo.narrativecraft.network.story.C2SStopStory;
 import fr.loudo.narrativecraft.platform.Services;
@@ -195,7 +193,7 @@ public class MainScreen extends Screen {
         if (!isPause) {
             leaveScreenButton = Button.builder(Translation.message("screen.main.leave_screen"), button -> {
                         minecraft.getSoundManager().stop(MAIN_MUSIC_INSTANCE);
-                        Services.PACKET.sendToServer(BiStopEditorMaker.INSTANCE);
+                        ClientNarrativeCraftMod.getInstance().getPlayerSession().requestEditorClose();
                     })
                     .bounds(width - BUTTON_WIDTH - 10, 10, BUTTON_WIDTH, BUTTON_HEIGHT)
                     .build();
@@ -252,13 +250,7 @@ public class MainScreen extends Screen {
     public void close() {
         minecraft.gui.setScreen(null);
         minecraft.getSoundManager().stop(MAIN_MUSIC_INSTANCE);
-        EditorMaker editor =
-                ClientNarrativeCraftMod.getInstance().getPlayerSession().getEditor();
-        if (editor != null) {
-            editor.stop();
-        }
-        ClientNarrativeCraftMod.getInstance().getPlayerSession().setEditor(null);
-        Services.PACKET.sendToServer(BiStopEditorMaker.INSTANCE);
+        ClientNarrativeCraftMod.getInstance().getPlayerSession().requestEditorClose();
     }
 
     @Override